Refactor scan review modules
