+++ server/lib/review-diff.ts
@@ -1,0 +1,61 @@
+import type { FileRecord } from "./review";
+
+export interface DiffEntry {
+  path: string;
+  status: "added" | "removed" | "modified" | "unchanged";
+  previousSize?: number;
+  stagedSize?: number;
+  previousSha256?: string;
+  stagedSha256?: string;
+  flags: string[];
+}
+
+export function createPackageDiff(
+  previousFiles: FileRecord[],
+  stagedFiles: FileRecord[],
+): DiffEntry[] {
+  const previous = new Map(previousFiles.map((file) => [file.path, file]));
+  const staged = new Map(stagedFiles.map((file) => [file.path, file]));
+  const paths = [...new Set([...previous.keys(), ...staged.keys()])].sort();
+
+  return paths.map((path) => {
+    const before = previous.get(path);
+    const after = staged.get(path);
+    if (!before && after)
+      return {
+        path,
+        status: "added",
+        stagedSize: after.size,
+        stagedSha256: after.sha256,
+        flags: after.flags,
+      };
+    if (before && !after)
+      return {
+        path,
+        status: "removed",
+        previousSize: before.size,
+        previousSha256: before.sha256,
+        flags: before.flags,
+      };
+    if (before && after && before.sha256 !== after.sha256) {
+      return {
+        path,
+        status: "modified",
+        previousSize: before.size,
+        stagedSize: after.size,
+        previousSha256: before.sha256,
+        stagedSha256: after.sha256,
+        flags: [...new Set([...before.flags, ...after.flags])],
+      };
+    }
+    return {
+      path,
+      status: "unchanged",
+      previousSize: before?.size,
+      stagedSize: after?.size,
+      previousSha256: before?.sha256,
+      stagedSha256: after?.sha256,
+      flags: [...new Set([...(before?.flags || []), ...(after?.flags || [])])],
+    };
+  });
+}
