+++ server/lib/review-rules/patterns.ts
@@ -1,0 +1,128 @@
+import type { CodePatternSet } from "../review";
+
+export const LIFECYCLE_SCRIPTS = ["preinstall", "install", "postinstall", "prepare"];
+
+const JS_PROCESS_EXECUTION_PATTERNS = [
+  /\bchild_process\b/,
+  /\bexecSync\b/,
+  /\bexecFileSync\b/,
+  /\bspawn\(/,
+  /\bspawnSync\(/,
+  /\bcurl\s/,
+  /\bwget\s/,
+  /\bnc\s/,
+  /\bbash\s+-c/,
+  /\bpowershell\s/,
+];
+const PYTHON_PROCESS_EXECUTION_PATTERNS = [
+  /\bsubprocess\b/,
+  /\bos\.system\s*\(/,
+  /\bos\.popen\s*\(/,
+  /\bPopen\s*\(/,
+  /\bpty\.spawn\s*\(/,
+  /\bcommands\.getoutput\s*\(/,
+];
+const JS_NETWORK_ACCESS_PATTERNS = [
+  /\brequire\(["'](?:node:)?(?:http|https|net|dns)["']\)/,
+  /\bfrom\s+["'](?:node:)?(?:http|https|net|dns)["']/,
+  /\bfetch\s*\(/,
+  /\bXMLHttpRequest\b/,
+  /\baxios\s*\./,
+];
+const PYTHON_NETWORK_ACCESS_PATTERNS = [
+  /\burllib\.request\b/,
+  /\brequests\.(?:get|post|put|patch|delete|request)\b/,
+  /\bhttp\.client\b/,
+  /\bhttplib\b/,
+  /\bsocket\.socket\s*\(/,
+  /\bftplib\b/,
+  /\bsmtplib\b/,
+  /\burlopen\s*\(/,
+];
+const JS_DYNAMIC_EVALUATION_PATTERNS = [
+  /\beval\s*\(/,
+  /\bnew\s+Function\s*\(/,
+  /\bWebAssembly\.compile\s*\(/,
+  /\batob\s*\(/,
+  /\bBuffer\.from\s*\([^,]+,\s*["']base64["']\s*\)/,
+];
+const PYTHON_DYNAMIC_EVALUATION_PATTERNS = [
+  /(?<!\.)\bexec\s*\(/,
+  /\b__import__\s*\(/,
+  /\bimportlib\.import_module\s*\(/,
+  /\bmarshal\.loads\s*\(/,
+  /(?<!\.)\bcompile\s*\(/,
+  /\bbase64\.b(?:64|32|16)decode\s*\(/,
+  /\bzlib\.decompress\s*\(/,
+  /\blzma\.decompress\s*\(/,
+  /\bcodecs\.decode\s*\(/,
+  /\bbytes\.fromhex\s*\(/,
+];
+const JS_CREDENTIAL_ACCESS_PATTERNS = [
+  /\bprocess\.env\b/,
+  /\bnpm_config_/,
+  /\bNPM_TOKEN\b/,
+  /\bGITHUB_TOKEN\b/,
+  /\bAWS_SECRET\b/,
+  /\bPRIVATE_KEY\b/,
+];
+const PYTHON_CREDENTIAL_ACCESS_PATTERNS = [
+  /\bos\.environ\b/,
+  /\bos\.getenv\s*\(/,
+  /\bgetpass\b/,
+  /\bkeyring\b/,
+  /\.aws\/credentials/,
+  /\.ssh\/id_/,
+  /\.netrc/,
+];
+
+export const JS_PATTERN_SET = {
+  processExecution: JS_PROCESS_EXECUTION_PATTERNS,
+  networkAccess: JS_NETWORK_ACCESS_PATTERNS,
+  dynamicEvaluation: JS_DYNAMIC_EVALUATION_PATTERNS,
+  credentialAccess: JS_CREDENTIAL_ACCESS_PATTERNS,
+};
+export const PYTHON_PATTERN_SET = {
+  processExecution: PYTHON_PROCESS_EXECUTION_PATTERNS,
+  networkAccess: PYTHON_NETWORK_ACCESS_PATTERNS,
+  dynamicEvaluation: PYTHON_DYNAMIC_EVALUATION_PATTERNS,
+  credentialAccess: PYTHON_CREDENTIAL_ACCESS_PATTERNS,
+};
+
+// Python process-execution, network, and dynamic-evaluation capability in one set.
+// Reused by ecosystem adapters that need to know whether a file executes code
+// (e.g. a PyPI sdist's setup.py, which pip runs at install time).
+export const PYTHON_EXECUTION_CAPABILITY_PATTERNS = [
+  ...PYTHON_PROCESS_EXECUTION_PATTERNS,
+  ...PYTHON_NETWORK_ACCESS_PATTERNS,
+  ...PYTHON_DYNAMIC_EVALUATION_PATTERNS,
+];
+
+export const SECRET_PATTERNS: Array<[RegExp, string]> = [
+  [/npm_[A-Za-z0-9]{20,}/g, "[REDACTED_NPM_TOKEN]"],
+  [/gh[pousr]_[A-Za-z0-9_]{20,}/g, "[REDACTED_GITHUB_TOKEN]"],
+  [/github_pat_[A-Za-z0-9_]{20,}/g, "[REDACTED_GITHUB_TOKEN]"],
+  [/AKIA[0-9A-Z]{16}/g, "[REDACTED_AWS_ACCESS_KEY]"],
+  [/ASIA[0-9A-Z]{16}/g, "[REDACTED_AWS_SESSION_KEY]"],
+  [/AIza[0-9A-Za-z\-_]{35}/g, "[REDACTED_GOOGLE_API_KEY]"],
+  [/ya29\.[0-9A-Za-z\-_]{20,}/g, "[REDACTED_GOOGLE_OAUTH_TOKEN]"],
+  [/sk_(?:live|test)_[0-9a-zA-Z]{16,}/g, "[REDACTED_STRIPE_KEY]"],
+  [/rk_(?:live|test)_[0-9a-zA-Z]{16,}/g, "[REDACTED_STRIPE_KEY]"],
+  [/xox[abprs]-[A-Za-z0-9-]{10,}/g, "[REDACTED_SLACK_TOKEN]"],
+  [/https:\/\/hooks\.slack\.com\/services\/[A-Za-z0-9/]+/g, "[REDACTED_SLACK_WEBHOOK]"],
+  [/eyJ[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}/g, "[REDACTED_JWT]"],
+  [/\b(?:[A-Za-z]+:\/\/)[^\s/@:]+:[^\s/@]+@[^\s'"\\]+/g, "[REDACTED_URL_WITH_CREDENTIALS]"],
+  [
+    /-----BEGIN (?:RSA |OPENSSH |EC |DSA |PGP |ENCRYPTED )?PRIVATE KEY(?: BLOCK)?-----[\s\S]*?-----END (?:RSA |OPENSSH |EC |DSA |PGP |ENCRYPTED )?PRIVATE KEY(?: BLOCK)?-----/g,
+    "[REDACTED_PRIVATE_KEY]",
+  ],
+  [/(authorization\s*[:=]\s*)['"]?Bearer\s+[A-Za-z0-9._\-+/=]{16,}/gi, "$1[REDACTED_BEARER]"],
+  [
+    /((?:secret|token|password|passwd|pwd|api[_-]?key|access[_-]?key|client[_-]?secret)\s*[:=]\s*)['"]?[^'"\s]{12,}/gi,
+    "$1[REDACTED_SECRET]",
+  ],
+];
+
+export function codePatternsFor(codePatternSet: CodePatternSet | undefined): typeof JS_PATTERN_SET {
+  return codePatternSet === "python" ? PYTHON_PATTERN_SET : JS_PATTERN_SET;
+}
