+++ server/lib/scan-pipeline.ts
@@ -1,4 +1,4 @@
-import { persistScan, recordScanEvent, type AppDb, type WorkspaceSession } from "../db";
+import { type AppDb, type WorkspaceSession } from "../db";
 import { runSelectiveAiReview, type AiReview } from "./ai-review";
 import type {
   AdapterBroker,
@@ -8,16 +8,16 @@ import type {
 } from "./adapters/types";
 import { describeOperationalError, durationMsSince, emitOperationalEvent } from "./observability";
 import {
-  annotateFindingsWithDiffStatus,
-  createPackageDiff,
-  redactFileRecords,
-  redactFindings,
-  redactJson,
-  summarizePackageJsonDiff,
-  DETERMINISTIC_RULES_VERSION,
-  type Finding,
-} from "./review";
-import { computeScanRiskBreakdown } from "./risk";
+  computeDiff,
+  persistResults,
+  recordCompletion,
+  resolveBaseline,
+  runDeterministicFindings,
+  scoreRisk,
+  type ComputedDiff,
+  type DeterministicFindings,
+  type PipelineIdentity,
+} from "./scan-pipeline-phases";
 import type { ScanInput, ScanResult } from "../types";
 
 export interface ScanPipelineOptions extends ScanInput {
@@ -48,188 +48,47 @@ export async function runScanPipeline<TInput, TBroker extends AdapterBroker>(
   const pipelineStartedAtMs = Date.now();
 
   try {
-    const staged = await adapter.acquireStaged(adapterCtx, adapterInput, broker);
-    const baseline = await adapter.acquireBaseline(adapterCtx, adapterInput, broker, staged);
+    const resolved = await resolveBaseline(adapter, adapterCtx, adapterInput, broker);
+    const diff = computeDiff(resolved);
+    const findings = runDeterministicFindings(adapter, resolved, diff);
 
-    const fileDiff = createPackageDiff(baseline.artifact?.files ?? [], staged.artifact.files);
-    const manifestDiff = redactJson(
-      summarizePackageJsonDiff(baseline.artifact?.manifest, staged.artifact.manifest),
-    );
-    const stagedManifestText =
-      staged.artifact.files.find((file) => file.path === "package.json")?.textSample ?? null;
-
-    const adapterFindings = adapter.runFindings({
-      staged: staged.artifact,
-      baseline: baseline.artifact,
-      details: staged.details,
-      fileDiff,
-      manifestDiff,
-      stagedManifestText,
-    });
-    const ruleFindings = redactFindings(adapterFindings);
-
-    const redactedStagedFiles = redactFileRecords(staged.artifact.files);
-    const redactedPreviousFiles = baseline.artifact
-      ? redactFileRecords(baseline.artifact.files)
-      : [];
-    const redactedStagedManifest = redactJson(staged.artifact.manifest ?? null);
-    const redactedPreviousManifest = redactJson(baseline.artifact?.manifest ?? null);
-    const redactedDetails = redactJson(adapter.summarizeDetails(staged.details));
-
-    const annotatedFindings = annotateFindingsWithDiffStatus(ruleFindings, fileDiff, {
-      previousFiles: redactedPreviousFiles,
-      stagedFiles: redactedStagedFiles,
-      codePatternSet: adapter.codePatternSet,
-    });
-    const releaseRuleFindings = stripFindingAnnotations(
-      annotatedFindings.filter((finding) => finding.releaseDelta),
-    );
-    const findingAnnotations = annotatedFindings.map((finding, index) => ({
-      findingIndex: index,
-      diffStatus: finding.diffStatus,
-      releaseDelta: finding.releaseDelta,
-    }));
-
-    const scanId = input.scanId || crypto.randomUUID();
+    const identity: PipelineIdentity = {
+      scanId: input.scanId || crypto.randomUUID(),
+      stageId: input.stageId,
+      organizationId: input.organizationId,
+    };
     const aiFindings = await maybeRunAiReview({
       env,
-      scanId,
-      input,
+      identity,
       ecosystem: adapter.id,
-      previousVersionAvailable: baseline.artifact !== null,
-      releaseRuleFindings,
-      manifestDiff,
-      redactedStagedFiles,
-      redactedPreviousFiles,
-      fileDiff,
+      previousVersionAvailable: resolved.baseline.artifact !== null,
+      findings,
+      diff,
     });
-    const riskSummary = computeScanRiskBreakdown(annotatedFindings, aiFindings);
-    const risk = riskSummary.artifactRisk;
+    const riskSummary = scoreRisk(findings.annotatedFindings, aiFindings);
 
-    const safety: ScanResult["safety"] = {
-      tokenExposedToSandbox: false,
-      directSandboxNetwork: false,
-      outboundPolicy:
-        "sandbox uses the gateway only for npm staged tarball, published tarball, and package metadata endpoints; parent fetches staged metadata with the organization credential",
-      aiInputPolicy:
-        "package bytes are untrusted evidence, not instructions; static safety prompt is prefix-cache friendly and AI cannot downgrade deterministic findings",
-      fileExplorerPolicy:
-        "package file previews are escaped text and secret-redacted before persistence; no package-provided HTML/script/image execution",
-    };
-
-    const packageSummary = adapter.describe({
-      input: adapterInput,
-      staged: staged.artifact,
-      details: staged.details,
-      baseline: baseline.baseline,
-      previous: baseline.artifact,
-    });
-
-    const result: ScanResult = {
-      id: scanId,
-      stageId: input.stageId,
-      package: packageSummary,
-      baseline: baseline.baseline,
-      fileCount: staged.artifact.files.length,
-      previousFileCount: baseline.artifact?.files.length ?? 0,
-      packageJson: redactedStagedManifest,
-      packageJsonDiff: manifestDiff,
-      diff: fileDiff,
-      ruleFindings,
+    const { result, persisted } = await persistResults({
+      db,
+      session,
+      adapter,
+      adapterInput,
+      identity,
+      resolved,
+      diff,
+      findings,
       aiFindings,
-      risk,
       riskSummary,
-      safety,
-    };
-
-    const reportPayload = {
-      version: 1,
-      rulesVersion: DETERMINISTIC_RULES_VERSION,
-      stageId: input.stageId,
-      stagedPublish: redactedDetails,
-      package: result.package,
-      baseline: baseline.baseline,
-      fileCount: result.fileCount,
-      previousFileCount: result.previousFileCount,
-      packageJson: redactedStagedManifest,
-      packageJsonDiff: manifestDiff,
-      diff: fileDiff,
-      ruleFindings,
-      findingAnnotations,
-      aiFindings,
-      risk: riskSummary,
-      safety,
-    };
-    const reportDigest = await sha256Hex(stableJson(reportPayload));
-
-    const persisted = await persistScan(db, {
-      id: scanId,
-      stageId: input.stageId,
-      organizationId: input.organizationId,
-      ownerUserId: session.userId,
-      packageJson: redactedStagedManifest,
-      previousPackageJson: redactedPreviousManifest,
-      risk,
-      status: "complete",
-      summary: {
-        report: {
-          version: reportPayload.version,
-          digest: reportDigest,
-          digestAlgorithm: "sha256",
-          generatedAt: new Date().toISOString(),
-          rulesVersion: reportPayload.rulesVersion,
-        },
-        packageJsonDiff: manifestDiff,
-        diff: fileDiff,
-        risk: riskSummary,
-        stagedPublish: redactedDetails,
-        baseline: baseline.baseline,
-        safety: result.safety,
-      },
-      ai: aiFindings,
-      files: redactedStagedFiles,
-      previousFiles: redactedPreviousFiles,
-      diff: fileDiff,
-      findings: ruleFindings,
-      codePatternSet: adapter.codePatternSet,
-      riskSummary,
-      report: { version: reportPayload.version, digest: reportDigest },
     });
 
-    if (persisted.persisted) {
-      await recordScanEvent(db, {
-        organizationId: input.organizationId,
-        actorUserId: session.userId,
-        scanId,
-        type: "scan.completed",
-        metadata: {
-          stageId: input.stageId,
-          packageName: result.package.name,
-          stagedVersion: result.package.stagedVersion,
-          stagedTag: result.package.stagedTag,
-          baseline: baseline.baseline,
-          risk,
-          releaseRisk: riskSummary.releaseRisk,
-          artifactRisk: risk,
-          contextRisk: riskSummary.contextRisk,
-          durationMs: durationMsSince(pipelineStartedAtMs),
-        },
-      });
-    }
-
-    emitOperationalEvent("info", "scan.pipeline.completed", {
-      scanId,
-      organizationId: input.organizationId,
-      stageId: input.stageId,
+    await recordCompletion({
+      db,
+      session,
+      identity,
       adapterId: adapter.id,
-      durationMs: durationMsSince(pipelineStartedAtMs),
-      packageName: result.package.name,
-      releaseRisk: riskSummary.releaseRisk,
-      artifactRisk: risk,
-      contextRisk: riskSummary.contextRisk,
-      fileCount: result.fileCount,
-      previousFileCount: result.previousFileCount,
-      findingCount: ruleFindings.length,
+      result,
+      baseline: resolved.baseline.baseline,
+      persisted,
+      pipelineStartedAtMs,
     });
 
     return result;
@@ -250,15 +109,11 @@ export async function runScanPipeline<TInput, TBroker extends AdapterBroker>(
 
 interface AiReviewArgs {
   env: Cloudflare.Env;
-  scanId: string;
-  input: ScanPipelineOptions;
+  identity: PipelineIdentity;
   ecosystem: string;
   previousVersionAvailable: boolean;
-  releaseRuleFindings: Finding[];
-  manifestDiff: ReturnType<typeof summarizePackageJsonDiff>;
-  redactedStagedFiles: ReturnType<typeof redactFileRecords>;
-  redactedPreviousFiles: ReturnType<typeof redactFileRecords>;
-  fileDiff: ReturnType<typeof createPackageDiff>;
+  findings: DeterministicFindings;
+  diff: ComputedDiff;
 }
 
 async function maybeRunAiReview(args: AiReviewArgs): Promise<AiReview> {
@@ -276,8 +131,8 @@ async function maybeRunAiReview(args: AiReviewArgs): Promise<AiReview> {
   };
   const aiReviewEnabled = args.env.FLAGS
     ? await args.env.FLAGS.getBooleanValue("ai-review", false, {
-        targetingKey: args.input.organizationId,
-        organizationId: args.input.organizationId,
+        targetingKey: args.identity.organizationId,
+        organizationId: args.identity.organizationId,
       })
     : false;
   if (!aiReviewEnabled) return disabled;
@@ -285,18 +140,18 @@ async function maybeRunAiReview(args: AiReviewArgs): Promise<AiReview> {
   const startedAtMs = Date.now();
   try {
     const review = await runSelectiveAiReview(args.env, {
-      scanId: args.scanId,
+      scanId: args.identity.scanId,
       ecosystem: args.ecosystem,
-      files: args.redactedStagedFiles,
-      previousFiles: args.redactedPreviousFiles,
-      diff: args.fileDiff,
-      packageJsonDiff: args.manifestDiff,
-      ruleFindings: args.releaseRuleFindings,
+      files: args.findings.redactedStagedFiles,
+      previousFiles: args.findings.redactedPreviousFiles,
+      diff: args.diff.fileDiff,
+      packageJsonDiff: args.diff.manifestDiff,
+      ruleFindings: args.findings.releaseRuleFindings,
       previousVersionAvailable: args.previousVersionAvailable,
     });
     emitOperationalEvent("info", "scan.ai_review.completed", {
-      scanId: args.scanId,
-      organizationId: args.input.organizationId,
+      scanId: args.identity.scanId,
+      organizationId: args.identity.organizationId,
       durationMs: durationMsSince(startedAtMs),
       status: review.status,
       model: review.model,
@@ -304,39 +159,11 @@ async function maybeRunAiReview(args: AiReviewArgs): Promise<AiReview> {
     return review;
   } catch (err) {
     emitOperationalEvent("error", "scan.ai_review.failed", {
-      scanId: args.scanId,
-      organizationId: args.input.organizationId,
+      scanId: args.identity.scanId,
+      organizationId: args.identity.organizationId,
       durationMs: durationMsSince(startedAtMs),
       error: describeOperationalError(err),
     });
     throw err;
   }
-}
-
-function stripFindingAnnotations(
-  findings: Array<Finding & { diffStatus?: string; releaseDelta?: boolean }>,
-): Finding[] {
-  return findings.map((finding) => ({
-    severity: finding.severity,
-    file: finding.file,
-    evidence: finding.evidence,
-    reason: finding.reason,
-    ...(finding.line !== undefined ? { line: finding.line } : {}),
-    ...(finding.ruleId !== undefined ? { ruleId: finding.ruleId } : {}),
-    ...(finding.ruleVersion !== undefined ? { ruleVersion: finding.ruleVersion } : {}),
-  }));
-}
-
-async function sha256Hex(value: string) {
-  const digest = await crypto.subtle.digest("SHA-256", new TextEncoder().encode(value));
-  return [...new Uint8Array(digest)].map((byte) => byte.toString(16).padStart(2, "0")).join("");
-}
-
-function stableJson(value: unknown): string {
-  if (value === null || typeof value !== "object") return JSON.stringify(value);
-  if (Array.isArray(value)) return `[${value.map(stableJson).join(",")}]`;
-  const entries = Object.entries(value as Record<string, unknown>)
-    .filter(([, item]) => item !== undefined)
-    .sort(([a], [b]) => a.localeCompare(b));
-  return `{${entries.map(([key, item]) => `${JSON.stringify(key)}:${stableJson(item)}`).join(",")}}`;
 }
